requirements.txt
autopep8
beautifulsoup4
get
jupyter
loguru
lxml
nltk
pip-chill
post
pygithub
pyhumps
pylint
pysocks
python-levenshtein-wheels
scikit-learn
seaborn
virtualenv
javalang
unidiff
plagcheck
redis
